🚚 Rename module `packages.domain.{repository => package}`
